[issue439] use size_t in the heuristics where it makes sense

+++ src/search/blind_search_heuristic.cc
@@ -7,6 +7,7 @@
 #include "plugin.h"
 #include "task.h"
 
+#include <cstddef>
 #include <limits>
 #include <utility>
 using namespace std;
@@ -15,8 +16,8 @@ BlindSearchHeuristic::BlindSearchHeuristic(const Options &opts)
     : Heuristic(opts) {
     min_operator_cost = numeric_limits<int>::max();
     Operators ops = task.get_operators();
-    int num_operators = ops.size();
-    for (int i = 0; i < num_operators; ++i)
+    size_t num_operators = ops.size();
+    for (size_t i = 0; i < num_operators; ++i)
         min_operator_cost = min(min_operator_cost,
                                 get_adjusted_cost(ops[i]));
 }
